// src/app/model/param-list.interface.ts
export interface ParamListInterface {
  // transactionDescription: string;
  secretCode?: string;
  /*tranDate: Date;
  sourceAccount: string;
  sourceAccountCurrency: string;
  sourceAccountType: string;*/
  senderName: string;
  senderMobileNo: string;
  amount: number;
  senderId?: string;
  beneficiaryMobileNo: string;
  // ccy: string;
  beneficiaryName: string;
  // withdrawalChannel: string;
}